Combine EntityAccessControlHandler and UnprocessableEntityAccessControlHandler into their base class
Also, do not apply `cachePerUser()` for the `$operation any` cases.

When `$operation any` is not allowed, it then checks the `$operation
own` permissions.
